+++ tests/integration/test_scaling_focus.gd
@@ -13,7 +13,7 @@ func test_single_card_focus_use_rectangle():
 	cfc.hand_use_oval_shape = false
 	cards[0]._on_Card_mouse_entered()
 	yield(yield_to(cards[0].get_node('Tween'), "tween_all_completed", 1), YIELD)
-	assert_almost_eq(Vector2(41.5, -240),cards[0].position,Vector2(2,2),
+	assert_almost_eq(Vector2(44.5, -240),cards[0].position,Vector2(2,2),
 			"Card dragged in correct global position")
 	assert_almost_eq(Vector2(1.5, 1.5),cards[0].scale,Vector2(0.1,0.1),
 			"Card has correct scale")
@@ -29,7 +29,7 @@ func test_single_card_focus_use_oval():
 	cfc.hand_use_oval_shape = true
 	cards[0]._on_Card_mouse_entered()
 	yield(yield_to(cards[0].get_node('Tween'), "tween_all_completed", 1), YIELD)
-	assert_almost_eq(Vector2(102.0, -213.5),cards[0].position,Vector2(2,2),
+	assert_almost_eq(Vector2(103.0, -240.0),cards[0].position,Vector2(2,2),
 			"Card dragged in correct global position")
 	assert_almost_eq(Vector2(1.5, 1.5),cards[0].scale,Vector2(0.1,0.1),
 			"Card has correct scale")
@@ -45,13 +45,13 @@ func test_card_focus_neighbour_push_use_rectangle():
 	cfc.hand_use_oval_shape = false
 	cards[2]._on_Card_mouse_entered()
 	yield(yield_for(1), YIELD)
-	assert_almost_eq(Vector2(23, 0),cards[0].position,Vector2(2,2),
+	assert_almost_eq(Vector2(25.75, 0),cards[0].position,Vector2(2,2),
 			"Card dragged in correct global position")
-	assert_almost_eq(Vector2(131.5, 0),cards[1].position,Vector2(2,2),
+	assert_almost_eq(Vector2(134.5, 0),cards[1].position,Vector2(2,2),
 			"Card dragged in correct global position")
-	assert_almost_eq(Vector2(686.5, 0),cards[3].position,Vector2(2,2),
+	assert_almost_eq(Vector2(689.5, 0),cards[3].position,Vector2(2,2),
 			"Card dragged in correct global position")
-	assert_almost_eq(Vector2(795.25, 0),cards[4].position,Vector2(2,2),
+	assert_almost_eq(Vector2(798.25, 0),cards[4].position,Vector2(2,2),
 			"Card dragged in correct global position")
 	cfc.hand_use_oval_shape = true
 func test_card_focus_neighbour_push_use_oval():
@@ -62,9 +62,9 @@ func test_card_focus_neighbour_push_use_oval():
 			"Card dragged in correct global position")
 	assert_almost_eq(Vector2(180.0, -40.0),cards[1].position,Vector2(2,2),
 			"Card dragged in correct global position")
-	assert_almost_eq(Vector2(638.5, -40.0),cards[3].position,Vector2(2,2),
+	assert_almost_eq(Vector2(643.5, -40.0),cards[3].position,Vector2(2,2),
 			"Card dragged in correct global position")
-	assert_almost_eq(Vector2(717.0, -22.5),cards[4].position,Vector2(2,2),
+	assert_almost_eq(Vector2(722.0, -22.5),cards[4].position,Vector2(2,2),
 			"Card dragged in correct global position")
 	cfc.hand_use_oval_shape = true
 func test_card_change_focus_to_neighbour():